+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: yooshima <yooshima@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/06 16:45:28 by yooshima          #+#    #+#             */
-/*   Updated: 2024/05/09 12:22:02 by yooshima         ###   ########.fr       */
+/*   Updated: 2024/05/09 16:11:58 by yooshima         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 
 int		ft_printf(const char *format, ...);
 int		if_per(const char *start, va_list ap, int count);
-int		put_hex(unsigned long long nb, int upper);
+int		put_hex(unsigned long nb, int upper);
 int		put_pointer(unsigned long long p);
 int		put_str(char *s);
 int		put_c(int c);
